Update patch set 1
Patch Set 1:

I looked at the Jira and it remains unclear what is this for?  I get that its some encrypted credential, but where is it used?  Is there some complementary change to a helm chart that needs to be made (to reference this new var)?  If so, does component chart owner need to make this change, or is it in some common chart?

Patch-set: 1