Update patch set 1
Patch Set 1: Code-Review+1

Looks good Ram.   

A side note, the info in NOTES.txt I don't think is necessary and information is incorrect (I think this file is common practice to be pasted around, so it is all over).

Another observation is that in an ONAP OOM deployment, the pod name will show up in a deployment as "dev-policy-policy-pap".   I know that the OOM team preferred the <component> prefixed to the pod name in the past, and I think that's still their preference but a side effect is the double policy in an official installation.

As a note we have "policy-pap" and "pap" components, we may have to explain folks the difference.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1