Update patch set 1
Patch Set 1: Code-Review+1

This is a great first contribution to Ingress. Thank you!

A necessary evolution of this work would be to introduce a global.ingress.enabled flag and an associated _ingress.tpl (e.g. common.ingress). The global flag can be added to the ONAP parent helm chart. The 'common.ingress' template would then be used at the top of an ingress.yaml to allow for global enablement of ingress for all helm charts - with ability to override with the ingress.enabled flag for any of the helm charts. I have raised a Jira to track this addition https://jira.onap.org/browse/OOM-1690. It is unassigned so please feel free to pick this up if so desired :)
Mike.

Patch-set: 1
Reviewer: Gerrit User 127 <127@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1