Improve override usability
Currently when you deploy onap with no override file, the behavior is
to deploy everything. In order to deploy a subset of components, an
override file must contain all components and then disable the ones
you don't want. As we prepare to transfer helm chart ownership to the
teams, it will simplify the creation of project specific development
override files, if the default behavior for deploying onap was
reversed. Allowing override files to only contain the components
they care to enable (and configure) and ignore the rest as they
would be disabled by default.

From this point on, it will be necessary to use an override file
(as integration uses for testing) to enable all components. This
patch includes an onap-all.yaml override file that may be used for
this purpose.

helm deploy dev local/onap -f onap/overrides/onap-all.yaml

No configuration is part of this override. Its purpose is only to
enable the components and is intended to be used in combination with
other override files that provide environment specific configuration.

Change-Id: I4b74a3a8a35a178298af7205762e2aca7c65dda3
Issue-ID: OOM-1692
Signed-off-by: Mike Elliott <mike.elliott@amdocs.com>