Update patch set 3
Patch Set 3:

Added some changes on top of Chenfei's trying to honor Borislav's comments in https://gerrit.onap.org/r/#/c/79102/:

policy-api is configured as a clusterIP service.   There's no initial intent to expose this port externally.  Nevertheless, the 30440 port (304 prefix is for ports that are not used at runtime unless enabled through configuration) is supported if we decide to make the switch to a NodePort in the future.

Patch-set: 3