Attempt2: Post-install jobs for DMaaP Provisioning
This was originally committed under 81992 but got a merge conflict.
I messed up resolving it so I abandoned 81992 and created this new
commit.

patchset2: replaced Release.Name for common.fullname in one of the
charts

patchset3: at Borislav's request, changed other references to
common.fullname so the approach is consistent throughout the
configmaps used for initialization.
Along the way, hit a limit of 53 chars on some resource name.
So changed the directory
from the verbose dmaap-bus-controller to the more terse dbc.

Change-Id: I1078003de094fee95c681f7fdaecb567235b7b3e
Signed-off-by: dglFromAtt <dgl@research.att.com>
Issue-ID: DMAAP-904
Signed-off-by: dglFromAtt <dgl@research.att.com>