Update patch set 2
Patch Set 2:

We already worked on this, we are waiting for AAF charts(not related to exposing Kafka NodePort) to be updated to submit a merge request.

Exposing Kafka Brokers outside the Kafka cluster is mandatory for deploying the mirror maker. More changes are required then the changes in this commit to expose the Kafka brokers outside the Kubernetes Cluster.

As we are planning to submit the changes after 03/22, I preferred to abandon these changes. You may refer the following K8S service ( need access to WindRiver) to see how we implemented the external access to the Kafka

http://10.12.7.5:8080/r/projects/1a264/kubernetes-dashboard:9090/#!/service?namespace=onap

Patch-set: 2