Update patch set 7
Patch Set 7:

Borislav, I am a little bit confused about the feedback with regards to when to allocate a NodePort or not.   I understood from your comment in this review:

https://gerrit.onap.org/r/#/c/81977/7/kubernetes/policy/charts/policy-xacml-pdp/templates/service.yaml@39

That the OOM team suggests to even if the intent is to use a ClusterIP, it's better to allocate a NodePort as a placeholder.   At the time, it seemed to me as an overkill, but we moved forward trying to follow this guideline.

I realize there may be ambiguity in your comment and/or misunderstandings in our part.

Can you clarify what's the recommendation to follow?

Patch-set: 7