Update patch set 3
Patch Set 3:

> (4 comments)

Gating job is just a full deployment of a fresh ONAP based on OOM Master overwritten by the tested patchest

for all the patchests the installation was completed.
once installation is finished, we take a snapshot of the pods and the helm charts, then run the healthcheck then run some End to End tests (if healthcheck core/small) are OK.

of course as the patchset includes the Master repo, it includes the current technical debt

that is why it gives just an indication that could be useful if you introduce a regression. 
But if the system was already broken and your patch is not aiming to fix that, the results will be a broken system for which your patchset has nothing to deal with..

typidically here, dmaap helm chart is not Completed for several days, but it has nothing to deal with the current patchset a priori

that is why we do not cast automatic vote and usually report the feedback with 0 

******** Kubernetes Results ********
 * Nb Pods: 199
 * Nb Failed Pods: 25
 * List of Failed Pods: [
   - onap-aai-aai-6d4b4bfb5-2c6kh
   - onap-aai-aai-champ-7f97f98976-qbdjx
   - onap-aai-aai-graphadmin-c79fb7fcf-szhtt
   - onap-aai-aai-graphadmin-create-db-schema-pzd4d
   - onap-aai-aai-resources-557c6f9d78-99c2x
   - onap-aai-aai-sparky-be-5ff8689c54-d4x5n
   - onap-aai-aai-spike-77ccc5bf78-mgznv
   - onap-aai-aai-traversal-7fb5558d98-w2cmv
   - onap-aai-aai-traversal-update-query-data-pnc84
   - onap-contrib-netbox-app-695fd89fd-k2kg7
   - onap-dcaegen2-dcae-bootstrap-d8bbc99f7-4k5hg
   - onap-dcaegen2-dcae-cloudify-manager-d7b597474-8ddnf
   - onap-dcaegen2-dcae-deployment-handler-5dcd748c6-n5nn4
   - onap-dcaegen2-dcae-healthcheck-84775fddd8-b5k68
   - onap-dcaegen2-dcae-policy-handler-85c5b8fcb9-5fmmh
   - onap-dmaap-dmaap-bc-68f4565db7-l2ldd
   - onap-sdnc-blueprints-processor-6d8c669cb-zvbt7
   - onap-sdnc-controller-blueprints-64b8cf4c7c-94xkx
   - onap-so-so-bpmn-infra-7b4cf7dbf4-8hqrt
   - onap-so-so-catalog-db-adapter-684f7f7f57-dfjx7
   - onap-so-so-f57fb6c78-7rbxv
   - onap-so-so-openstack-adapter-6c75fcdd89-jlgtb
   - onap-so-so-request-db-adapter-6797d64c57-xpkng
   - onap-so-so-sdc-controller-849949db55-8vgrs
   - onap-so-so-vfc-adapter-6df5476d68-wddjc]
 * Deployment Duration Estimation (s): 3600

*********** Helm Results ***********
 * Nb Helm Charts: 26
 * Nb Failed Helm Charts: 2
 * List of Failed Helm Charts: [
   - onap-dmaap
   - onap-vnfsdk]

helm status onap-dmaap -o yaml
info:
  Description: 'Release "onap-dmaap" failed post-install: timed out waiting for the
    condition'
  first_deployed:
    nanos: 379767395
    seconds: 1553250572
  last_deployed:
    nanos: 379767395
    seconds: 1553250572
  status:
    code: 4
name: onap-dmaap
namespace: onap



Basic DMAAP Data Router Health Check                       | PASS |
----------------------------------------------------------------------
Basic DMAAP Message Router Health Check                    | PASS |
----------------------------------------------------------------------

Patch-set: 3
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0