Update patch set 4
Patch Set 4:

I think there's a circular dependency here.  The root aai pod waits for traversal and resources to be up before init, but those pods need the pem file that the root pod mounts for them.  In my environment the aai pod never comes up and traversal and resources are in CrashLoopBackoff

Patch-set: 4