Update patch set 3
Patch Set 3:

In my tests, executing the DMaaP post install hooks is completing in 5 minutes(default time out). If there are some performance issues it may take more time. I heard from the OOM team that AAI is also going to add some post install hooks and they are planning to update the default time out in deploy plugin. 

Time taken for the start of the Mirror Maker will not have any impact on any container. Creation of the AAF permissions(through one of the DMaaP  post install hook)and updation of cache will take some time. If starting of any application depend on the AAF permissions and the permissions are created at run time , it will definitely take some time to complete. It is not an issue with the DMaaP


We requested the AAF team to update the AAF DB with the data from the AAF TEST, so that these permissions will be available immediately after the AAF deployment. Once the AAF updated their charts with the data from the AAF TEST, deployment of DMaaP mirror maker will not take significant time.

Patch-set: 3
Reviewer: Gerrit User 2343 <2343@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0