Update patch set 1
Patch Set 1:

In the actual (since yesterday) sdnc gerrit Dockerfile the odl-netconf-connector-all was already removed. But it was in the container in the previouse version. 
My tests with the SDNC container with the gerrit/source was successfull and there was always only one netconf-topology active. I verified this. 
Perhaps the scipt should handle also removing of odl-netconf-connector-all.

Patch-set: 1