Update patch set 1
Patch Set 1:

If you install both odl-netconf-connector-all(which contains odl-netconf-topology) and odk-netconf-clustered-topology(which contains odl-netconf-topology as well).That now adds the plugin two times and causes the netconf clustering to dail as has been tested.
Please make sure odl-netconf-connector-all is not present in the final image.
Odl documentation clearly states that you cannot have both the features and this will cause the netconf clustering to fail

Patch-set: 1