Update patch set 2
Patch Set 2:

We (with Morgan) are just summarising what jobdeployer is showing. 
I do understand your point but we try to "force" module to be clean all the time.
As I think also your code is not the issue but somewhere else in APPC, I'm not putting -1. 
But as I cannot be 100% sure, I cannot also put +1.

As we are just putting +1/0/-1 and not allowed to +2/merge, it's not a big deal not to have the +1 from us IMHO.

I imagine that persons with +2/merge will accept the patch (that's what I would do) and my score of "0" is more to make an emphasis to APPC PTL (so you) that in our current deployment we have issues with APPC.

Patch-set: 2
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0