Update patch set 2
Patch Set 2: Code-Review-1

This conflicts with the galera change.  Once that one is merged, I'll resolve the merge conflict in this one.  I tried to rebase this change against that one, but when I tried to submit the review I saw a ton of other dependent commits so I backed off and thought I'll just wait to resolve the conflict when I can just rebase against master.

Patch-set: 2
Label: Code-Review=-1