Update patch set 1
Patch Set 1:

> could it be related to https://issues.apache.org/jira/browse/CASSANDRA-14475
 > ?

The patch is to fix the mismatch of attribute and its value in template. The actual value shown in the results is set from integration-override.yaml, which is part of Integration deployment parameter we exercise at deployment time. 

I have submitted the test results above, and the actual value of timeout is out of scope this patch. Please approve.

Patch-set: 1