Update patch set 2
Patch Set 2: Code-Review+1

title is a bit weird, isn't it?
It is more a patch to update OOM docker versions (aligned with manifest) to latest versions for Dublin.

Each team should have the reflex to do it (the pain has to be shared and cannot be only on integration team :)), the release manager could check the consistency (especially after M4)

as a results in gating, without surprise it leads to some changes but no dramatic changes

Gating is now closed to intergation daily test 

http://onapci.org/logs/windriver-rke-staging-daily/13/report.html#tags?health
healthcheck: 61 tests 53 OK / 8 Fail versus 61/50/11 for the gating 

Healthcheck Fail both in gating and integration daily
- Basic DMAAP Bus Controller Health Check With Basic Auth
- Basic Multicloud-fcaps API Health Check
- Basic Multicloud API Health Check
- Basic MUSIC Cassa Check
- Basic MUSIC Health Check
- Basic VFC jujuvnfmdriver API Health Check

healthcheck Fail in integration (but not in gating)
- Basic OOF-CMSO Health Check
- Basic VNFSDK Health Check

healthcheck Fail in gating (but not in integration)
- Basic AAF Health Check 
- Basic AAF SMS Health Check 
- Basic DCAE Health Check
- Basic Holmes Rule Management API Health Check   
- Basic Holmes Engine Management API Health Check 

more pods are not starting properly (dmaap, sdc, aaf, pomba,..) after the upgrade of the version, only dmaap helm chart in in fail status (was already the case in previous gating taks before the version updates) 

I would recommend to merge (and try to keep as sync as possible - in theory after M4 it should be always sync, no?) then create the JIRA accordingly for the pods/helm chart not running/healthcheck tests that do not work

Patch-set: 2
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1