Update patch set 2
Patch Set 2:

So I'm confused - my request was to update both parts of oom - the policy references to https and the robot references to policy's protocol and port. Its the same oom repo. Unless you are saying there will be new parameters required and the current https, port 8181, 6969 are valid but not applicable to the new API ?

Patch-set: 2