Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> the AAI charts have transitioned to a submodule. is the Orange deployer cloning/initialising the submodules of OOM?

Orange depleployed is including the submodules but there is an issue when changes are pushed into gerrit prior to the introduction of the submodules on dev side. Our gerrit review -d <gerrit_id> will systematically fail (there will be a git conflict).

Jessica from LF had a look. Her conclusion is "What devs need to do is to "git reset HEAD kubernetes/aai/" and proceed with "git commit" and push their change and it will be rebased into the latest tip. "

Patch-set: 6
Label: Code-Review=0