Update patch set 6
Patch Set 6: Verified-1

(2 comments)

In addition to the 2 comments, can I please ask that test results be added as a comment to this patch. We are looking for proof that the Helm Chart is successfully deployed and its pods are coming up successfully in an ONAP deployment.
Thanks,
Mike.

Patch-set: 6
Label: Verified=-1