Update patch set 3
Patch Set 3:

This patch address the expiring DMaaP MR certificate issue. 

We have seen this error before and this is due to the missing AAF permissions. AAF charts are updated on 05/06 and addresses the permissions issue. We deployed the DMaaP and AAF after the updated AAF charts are released and we dont see any issue. We are not sure why the latest AAF changes are not reflected in the deployment mentioned in the comments.

https://gerrit.onap.org/r/gitweb?p=oom.git;a=commit;h=d50ad92651e96b9645e98f9160cac156ed1050b7

Patch-set: 3
Reviewer: Gerrit User 2343 <2343@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0