Update patch set 7
Patch Set 7:

Hi Morgan, This commit is for the OSDF repo within the oof project, which is passing the health check (please see Basic OOF-SNIRO Health Check in the Gitlab Orange 2 run at 12:34 pm on patchiest 7). The homing health check is for the HAS within the OOF project, which should pass the health check once the following commit is merged (https://gerrit.onap.org/r/#/c/86209/). Given that SNIRO health check is looking good, can we go ahead and merge this commit ?

Patch-set: 7