Update patch set 8
Patch Set 8:

Hi Sylvain, This commit is for the OSDF repo within the oof project, which is passing the health check. The homing health check is for the HAS within the OOF project, which should pass the health check once the following commit is merged (https://gerrit.onap.org/r/#/c/86209/). Similarly, https://gerrit.onap.org/r/#/c/86672/ is for the CMSO and should be fixing the CMSO health check. Given that SNIRO health check is looking good, can we go ahead and merge this commit ?

Patch-set: 8
Label: Code-Review=0