Update patch set 1
Patch Set 1:

As far as I can tell this is a timing/load issue on portal db startup, and changing it back to 2.5.0-STAGING-latest will not fix the problem; there's also been no change in portal repos since 2.5.0 was released, so 2.5.0-STAGING-latest should have identical content as 2.5.0.

Portal team will need to release 2.5.1 if they want to try a fix.  In the mean time, teams that want to test with the staging version should override the portal versions with the next staging version 2.5.1-STAGING-latest.

Patch-set: 1