Update patch set 1
Patch Set 1:

last test on Master daily (included the extension of the timer) lead to a 502 (last Robot healthcheck failure)

But I need to check the override.yaml to see if the component is properly deployed on our side. I do not see parser pod in the list of 277 running pods.

as we got a 502 I assume something was answering but I cannot find the pod....
if your healthcheck calls a front end (msb) that calls your pod that is not started, it could explain the 502...

could you share your override.yaml for OOM installation?

Patch-set: 1