Fix issue with etcd pod startup
Etcd pod fails to start with file not found error.
The error is being caused because the pod is stuck
in an infinite loop as the readiness probe is not true.
Since, we are anyway checking the pod status, we removed
the readiness probe as it is not needed.
Bumped up version of etcd to 3.2.24
This fixes a known issue:
https://github.com/etcd-io/etcd/pull/4861

Issue-ID: MULTICLOUD-660
Change-Id: I815766b4a8f187d88bb2fcdb71e9d6e24b277d25
Signed-off-by: Kiran Kamineni <kiran.k.kamineni@intel.com>