Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> pods and charts OK but robot healthcheck KO
> Basic SDNC Health Check                                               | FAIL |
> Resolving variable '${resp.json()['output']['response-code']}' failed: ValueError: No JSON object could be decoded

This could be related to an issue I found in local testing.  'make all' is failing in
multicloud, causing the rest of the make to bail - so sdnc and cds helm charts are not
updated.  I was able to get around this in my local testing by explicitly doing
make sdnc and make cds, but really the multicloud make needs to be fixed.

Patch-set: 2
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0