Update patch set 2
Patch Set 2:

I think it matters to make database portable not to be case agnostic, otherwise porting or sharing databases in the future with other components will be challenging, if we go to common db resources.   I would like to discuss a little more on the chat some other aspects.

Patch-set: 2