Fix for appc clustering install failure
Update appc karaf stop detection
The new stop detection script is more reliable and
does not depend on a specific value and format of
the ps command.

Remove clustering feature install
The feature needs to be installed before appc features
are installed, and so the odl-mdsal-clustering install
has been moved to the build scripts.

Change-Id: I6e9c585cbd58a516114a609329422cf8375e023d
Signed-off-by: Patrick Brady <patrick.brady@att.com>
Issue-ID: APPC-1242