Update patch set 1
Patch Set 1: Code-Review+1

This change is needed to fix a bug introduced recently in the cds-ui helm charts.  2 values being passed as env settings are interpreted as non-strings (one as a Boolean, the other as an int) and helm deploy on sdnc is failing because the value is missing quotes.  I tested this fix in my WindRiver env and verified that it fixes the problem

Patch-set: 1
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1