Update patch set 1
Patch Set 1:

Yes, it's nothing weird. This commit is supposed to fail in this case. 

Think about it. The config is set to follow current branch in subproject. You have just created a new one (locally) for example ABC. So git tries to find branch ABC also in subproject repo but it obviously does not exist.

Patch-set: 1