Update patch set 1
Patch Set 1:

Ok, I investigated this issue a little bit more and here is the output:

1) git submodule update --init --recursive --remote

This command behaves correctly. It actually should fail in this case because after you init the submodule a checkout using commit hash is performed. So technically the repo is in a detached head state. Additionally according to the doc the default update strategy is checkout so it would also leave the repo in detached head.

2) git submodule init

This is the source of all the issues. The problem is that this command always clones the repo and leaves it in detached head state which causes for example the --rebase strategy of update command to not work out of the box as described in the doc. I consider this to be a bug or at least lack of some functionality so I'll report this to the git community

3) Temporary solution

It is actually possible to workaround all those issues if you prepare your repo correctly and then use --rebase strategy for submodule update. I made a small script which clones the doc repo and sets up everything as required:

 #!/bin/bash
 
 function prepare_submodules {
 	git submodule update --init
 	local SUBMODULE_LIST=`git submodule--helper status | awk '{print $2}'`
 	while read -r submod; do
 		local PARENT_DIR=`pwd`
 		local BRANCH=`git submodule--helper remote-branch ${submod}`
 
 		cd ${submod}
 		git branch -f ${BRANCH} HEAD;
 		git checkout ${BRANCH}
 		if [ ! -z "`git submodule status`" ]; then
 			prepare_submodules;
 		fi
 		cd ${PARENT_DIR}
 	done <<< "$SUBMODULE_LIST"
 }
 
 git clone ssh://kopasiak@gerrit.onap.org:29418/doc
 cd doc
 prepare_submodules

After you clone the repo with this script you get all submodules recursively initialized and all checked out to the local branch with a suitable name so after that you can use simply:

 git submodule update --remote --rebase --recursive

and there is no need to change anything inside the repo itself thus I'd prefer to leave this -1 as is.

Patch-set: 1