Move Storage access to RWO
Today when deploying NBI with OOM, the PersistentVolumeClaim needs the
"ReadWriteMany" (or "RWX") capability.

According to Kubernetes Documentation (https://kubernetes.io/docs/concepts/storage/persistent-volumes/#access-modes),
ReadWriteMany stands for "the volume can be mounted as read-write by many nodes".

That means that a particular PVC needs to be read and written from many pods.
That also means that your code takes that into account and do the work to avoid
write at the same place at the same time.

An issue on RWX mode is that most "official" storage driver from Kubernetes doesn't
support it (13 over the 19 drivers doesn't support it, espacially OpenStack, Amazon
and Google storage classes).

MongoDb PVC is used only for one Mongo instance and it's the same for
maria. Thus we don't need RWX

Change-Id: I4d54cd7853778d8ba6c83551a211abe34d383382
Issue-ID: EXTAPI-305
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>