Move SDC Storage to RWO
Today when deploying SDC with OOM, the PersistentVolumeClaim needs the
"ReadWriteMany" (or "RWX") capability.

According to Kubernetes Documentation (https://kubernetes.io/docs/concepts/storage/persistent-volumes/#access-modes),
ReadWriteMany stands for "the volume can be mounted as read-write by many nodes".

That means that a particular PVC needs to be read and written from many pods.
That also means that your code takes that into account and do the work to avoid
write at the same place at the same time.

An issue on RWX mode is that most "official" storage driver from Kubernetes doesn't
support it (13 over the 19 drivers doesn't support it, espacially OpenStack, Amazon
and Google storage classes).

SDC PVC for ES is used only for one POD.
Thus we don't need RWX.

The other values are not used as of today but changing them to RWO seems
better if they're reused again

Change-Id: I3dabbce120a14d19499330b51e808c5725588e0c
Issue-ID: SDC-2549
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>