Update patch set 1
Patch Set 1:

value in sac-onboarding-be is not used. 
It's the one in cert.persistence.accessMode which is used (and its 'ReadOnlyMany' so it can be used by many PODS).

So I changed it in case someone wants to use it (it would break the deployment on a public cloud).

For the staging vs master, I understood that staging was for testing "STAGING" images. As it's not an image change, I believe the place is in master.

@Brian/Mike, can you tell me if I'm wrong?

Patch-set: 1