Update patch set 1
Patch Set 1:

As I wrote previously I don't have any ultimate voice here. I'm just sharing my humble opinion/advisory and I'm perfectly fine if PTL/committer/maintainer disagree with me and merge this. That's the beauty of open source right?;) People may have different opinions/feeling on some piece of code but in the end of the day it's maintainer's ultimate decision whether merge the code because he is the one who is going to maintain it.

In my environment I just walked around this issue with the clone shell snippet that I provided in some earlier comment and jira ticket, without any source code modification. I've reported this also as an issue to git mailing list but there is no response so I'm not sure if it's gonna be fixed soon.

In general I don't support this approach but if PTL/committer/maintainer likes it feel free to merge it.

Patch-set: 1