Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> oh, right, I depend on the latest drools snapshot image, and "master" goes by released versions and "staging" that uses snapshot is no supported by oom team I understand.
> 
> I'll cherry pick to staging and run oom redeploy there to check.   Does it sound as an ok idea?   For this review then, I will have to put it on hold until we have a new released image for pdpd-cl.


or you change the container tag values here to STAGING to verify it works (you'll have a -1 for Jobbuilder but OOM deployment will be triggered). So you'll gain confidence that the staging versions are OK.
When they're released, you change the container tags value to the released one.
Nobody is checking "staging" except Gating (but Gating is also checking master)

Patch-set: 2