Update patch set 1
Patch Set 1:

No dependencies with music or any other components to be released for this review ..

We also need for integration testing this one to be merged at your earliest convenience:   

https://gerrit.onap.org/r/#/c/oom/+/94930/.   

No dependencies with others with this other one neither.

For the integration testing, core use cases: vFW, vDNS, .. 

we need this two above mentioned made available in labs.

Patch-set: 1