Disable AAF for PDP-D telemetry API
As AAF cadi libraries and Aether libraries (transitively
brought in by drools kie-ci functionality seem
to interfere with each other, AAF is disabled.

There is no side effects to disable AAF for this API
as it is not an external API (it is used for diagnostics
only).

The nexus value change is cosmetic, not really used.

Issue-ID: POLICY-2109
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>
Change-Id: I2ccb3b0f08a4d3020e58ca83bd89d2f614a51dca