Update patch set 4
Patch Set 4: Code-Review+1

gating is OK now - the issue was kubernetes but not ONAP, sorry for the inconvenience.

I saw that the verify job is failed. As far as I understand this is due to the recent change done by LF on a script used in the verify jjb
this is a python3 issue, probably not the last one :)

The script can be found here https://github.com/lfit/releng-global-jjb/blob/eeeb5656c1c3aec9ad08918b916263c81ec626b9/shell/python-tools-install.sh
it is included as submodule in ci-management

I suspect a side effect of the patch from the 24th of September (I can see it on github mirror, but no access to lfit gerrit review history) done by LF
to be confirmed by Aric.

You should raise a critical ticket to LF IT...as it is preventing to the publish helm charts :)

Patch-set: 4
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1