Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Christophe - can you test this patchset in your azure setup ?

Hi Brian,
As I have mentioned in my comments earlier, I need to submit a set of changes of the override files. 
I'll submit these change shortly. The idea of this submission is that everyone should be able to see the changes which are coming in.
This, combined with https://gerrit.onap.org/r/c/so/+/92631 changes, should be tested. The last time I checked AAF Services were down.

Patch-set: 4
Label: Code-Review=0