Update patch set 1
Patch Set 1:

Hello,
I do not see the ability to optionally disable AAF integration. This is expected as part of the M# security checklist. Other applications such as AAI, CLAMP already have a configuration value to do this. Since this does not yet exist in this helm chart please refer to  https://jira.onap.org/browse/OOM-1986 for the general global configuration value we would like to align all projects to.

Thanks,

Patch-set: 1