Update patch set 2
Patch Set 2:

Just documenting here what I responded previously in email.  As far as I can tell, the helm deploy uses the master branch version of this repo so until this is merged, not all use cases work.  If this is true, then I would recommend merging this commit and deal with any issues in a separate Jira.   ALso, the failures reported by the Orange JobBuilder are for other non-DCAE components, so they are not applicable to this commit.

Patch-set: 2