Update patch set 1
Patch Set 1:

Hi Sylvain, it was considered that would be good to have if there was  a new el alto maintenance release made available, that's the only reason, as the most recent version was not correct in el alto charts.  Nevertheless, not aware of side effects with current settings, although it would be best just in case to point to the right versions in El Alto.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0