Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> HC + Pods are OK but isn't a bit too late for El Alto?
> 
> What is the rationale of this (https://jira.onap.org/browse/POLICY-2171 is not really helpful...)?

Not too late - the component is our legacy component and it was not part of Integration testing. Subsequent testing revealed that these fields were not updated. Teams that were still using legacy will need this. As Jorge indicated, its best to point to the correct released versions.

Patch-set: 1