Update patch set 3
Patch Set 3:

I agree with Krzysztof that we better not add more work for the secrets task effort.
On the other hand, integrating with AAF to automatically generate a certificate is not trivial, and looks like we don't have the resources to do that.

Ilana, can you move the passwords to secrets instead? I think that's the best option for now.

Patch-set: 3
Label: Verified=0