Update patch set 1
Patch Set 1:

One question Chenfei, in the OOM side, they are still using the El Alto version.   Does this configuration would cause an issue with the  official El Alto version trying to preload the policy types.   If so, would also add the change to change the values.yaml with the snapshot version.   

In summary, I know that these changes are needed when testing with the snapshot image (which is not there), but not sure if they are backwards compatible with the official version in El Alto which the oom repo has, in which case, it will have to be changed to the snapshot-latest one.

Patch-set: 1