Update patch set 3
Patch Set 3:

@Sylvain
I absolutely agree with you and if it would be only me to decide about that I would not expose it.

Unfortunately this is a little bit more complicated as there is quite large variety of different deployment options and use cases and what we would like to avoid is making some of them stop working. That's why we are targeting to improve the situation step by step. First switch everything what's possible to use ingress to make sure that we have all of NodePort capabilities. Then in next release we will try to remove the usage of NodePorts and in the same time review what is exposed outside via ingress and remove whatever can be removed.

Does it make sense to you?

Patch-set: 3
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0