Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What is PV and PVC?

PV is Persistent Volume. When we deploy ONAP, as we are using an NFS share, we must create a folder in this folder for the datas we need to keep. When we'll move to dynamic PV, it will be created automatically by the storage class 
PVC is Persistent Volume Claim. This is the way to bind a "PV" into deployment or statefulset.

Patch-set: 1