Update patch set 6
Patch Set 6: Code-Review-1

Sylvain - we are not intending to merge this soon. It will be marked WIP. These configuration changes will not work with the El Alto release images. Until we are ready to release images for Frankfurt, we are queueing up the configuration changes so we don't lose track as we go along.

Patch-set: 6
Label: Code-Review=-1