Update patch set 6
Patch Set 6: Code-Review-1

Right, thanks a lot Sylvain for taking a look, and will do the changes you suggested on your comments.   So, as Pam said, this would be a running inspection until we release the first set of artifacts for Frankfurt.   I still need to do testing with this patch, but wanted to get the policy folks look at it first, since it integrates several of the changes being introduced in Frankfurt.   These changes need of the unreleased image version (the snapshot).   This is the reason to keep it as a running inspection till we release, because of the restriction in the repo to not submit code changes with snapshot images.   After folks take an initial look, I'll move into WIP so you don't see it in your queue of reviews.   Makes sense?   If you have better ideas on how to approach it, let me/us know.

Patch-set: 6
Label: Code-Review=-1