Update patch set 6
Patch Set 6:

Pamela, Jorge, Yes I have understood that this is work in progress. I was just saying that IMHO it was the good way to proceed with OOM :)

If you need to "regate" it once in a while just write oom*_redeploy (without the *) and it can give you an idea of how it work.

And when integration will add automated tests around policy, you'll also have direct feedback on the code also

Patch-set: 6
Label: Verified=0