Update patch set 3
Patch Set 3: Code-Review-1

I still think that we shouldn’t see the aaf value especially in the global part. 
You should IMHO create an option in the config part to set the aaf endpoint (as an fqdn) and not set it per default. 

(Semi)hard coding the address will fool everybody blindly deploying your chart. 

And if your aaf server is out of the cluster, you set the option with the fqdn

Patch-set: 3
Label: Code-Review=-1