Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> Comments/Questions:
> 
> There are some changes that may not at all related to Service Mesh changes(Some of them are trivial such as changing the port names and service name etc, but some will definitely create issues such as changing the Kafka endpoints in MsgRtrApi.properties). Please try to avoid them as much as possible

Changing the port names is mandatory in order for service mesh to know the type of traffic. I'm not sure I changed the name of a service but I definitely tried to normalize service creation.

> 
> In order to stop using AAF, more changes are required both in the Helm charts( disable the SASL related properties in server.properties) and Docker images. Is there any env variable based on which I can add and remove the AAF related code in Docker images

There are no such env variable in order to do that but it's fairly trivial to add. So please give me the name you've decided and I'll adapt the charts

> 
> Is there any documentation on what you are trying to do and its impact on the each application. What is the timeline for implementing this change (Looks like we need some changes in Docker images and as we are moving to the Confluent Platform, so we are completely rewriting the Kafka and Zookeeper images and  Helm charts, I have to include these change in the new charts)

Plan is to have a demo of instantiation of a service during Prague DDF. I can help on the helm chart (and docker also i believe) if you want

Patch-set: 9